Fix issue with Mac OS writing config.json to the users directory
Because pkg, which is used for deploying quotecord, creates its own
filesystem and bundles all require() files into the binary, the real
file system is accessed using process.execPath. For this same reason,
config.json is read using readFile, rather than simply require(), as
the file would otherwise be packaged into the distributable with no way
for the user to edit or access its contents.